Change the mode "informative" to atLeastTwo
The values selection mode used just before to compute distances can be
"union", "intersection" or "informative". However, "informative" is
used for qualifying traits that hold phylogenetic information (which
is a more correct of the word).

This commit change the "informative" selection mode to "atLeastTwo" to
make things clearer and more correct.